rust - use permutation iterator for graph
When iterating over which start and goal nodes to use for the graph
search, before this commit we would choose 100 with a weighted random
selection to prefer nodes earlier in each list because those were the
most frequently seen words.

With this commit we use the `permutation-iterator` library for iterating
over a random permutation of start and goal nodes. Moreover we also
attempt the first 2 starts and first 2 goals if present.

According to the benchmarks this makes the password generation more
expensive in average and more variable. This is as expected, but could
be a worse customer experience.