README - fix cargo build to avoid bench
The Rust library comes with benchmarks in the "bench" subfolder. However
Rust "bench" is a nightly-only feature, which means that you need the
cutting-edge version of Rust to use it, whereas for typically work you
stick to the "stable" version.

By doing the default `cargo build` you avoid executing the benchmarks
and builds are successful. This is actually what the Gradle file already
does, and the README file mistakenly did not.