build.gradle - Windows fixes, fix running Android
When running commands on Windows from Gradle you need to prefix them
with `cmd /c` or else sometimes they don't work. I found a StackOverflow
post which creates a wrapper method that will do this for Windows, and
not do this for other operating systems.

This commit also adds some commands for running the app on Android from
the command-line. This should also work for integration tests but I
didn't test that.